Gather all video-related types in a dedicated file
