+++ src/pages/work/[slug].tsx
@@ -4,7 +4,8 @@ import ProjectOptionalBlocks from "@/components/blocks/projectOptionalBlocks";
 import Grid from "@/components/layout/grid";
 import MetaTags from "@/components/layout/metaTags";
 import Wrapper from "@/components/layout/wrapper";
-import { ProjectProps, VideoLinkObject } from "@/types/interfaces";
+import { ProjectProps } from "@/types/interfaces";
+import { VideoLinkObject } from "@/types/video";
 import { getGlobalData, getProjectData } from "@/utils/dataQueries";
 import { getMainVideoDirectLinks } from "@/utils/vimeoQueries";
 import { GetStaticPaths, GetStaticProps } from "next";
